Move all reports to their own package.
