+++ sigridci/sigridci/reports/gitlab_pull_request_report.py
@@ -17,9 +17,9 @@ import os
 import urllib.error
 import urllib.request
 
-from .publish_options import RunMode
 from .report import Report
-from .upload_log import UploadLog
+from ..publish_options import RunMode
+from ..upload_log import UploadLog
 
 
 class GitLabPullRequestReport(Report):
