+++ sigridci/sigridci/reports/pipeline_summary_report.py
@@ -15,9 +15,9 @@
 import sys
 
 from .ascii_art_report import AsciiArtReport
-from .objective import Objective, ObjectiveStatus
-from .publish_options import RunMode
 from .report import Report
+from ..objective import Objective, ObjectiveStatus
+from ..publish_options import RunMode
 
 
 class PipelineSummaryReport(Report):
